[Chore] :: di module data module로 이전
