--- core/database/src/main/java/com/onmi/database/di/DataStoreModule.kt
@@ -1,29 +1,0 @@
-package com.onmi.database.di
-
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.PreferenceDataStoreFactory
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.preferencesDataStoreFile
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object DataStoreModule {
-    @Singleton
-    @Provides
-    fun providePreferencesDataStore(@ApplicationContext context: Context): DataStore<Preferences> {
-        return PreferenceDataStoreFactory.create(
-            migrations = emptyList(),
-            scope = CoroutineScope(Dispatchers.IO),
-            produceFile = { context.preferencesDataStoreFile("onmi.datastore") }
-        )
-    }
-}
