+++ data/src/main/java/com/onmi/data/di/DataModule.kt
@@ -1,11 +1,15 @@
-package khs.onmi.aos.modules
+package com.onmi.data.di
 
+import com.onmi.data.datasource.LocalOptionDataSource
+import com.onmi.data.datasource.LocalUserDataSource
 import com.onmi.data.datasourceimpl.MealDataSourceImpl
 import com.onmi.data.datasourceimpl.SchoolDataSourceImpl
 import com.onmi.data.datasourceimpl.TimeTableDataSourceImpl
 import com.onmi.data.datasource.MealDataSource
 import com.onmi.data.datasource.SchoolDataSource
 import com.onmi.data.datasource.TimeTableDataSource
+import com.onmi.data.datasourceimpl.LocalOptionDataSourceImpl
+import com.onmi.data.datasourceimpl.LocalUserDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -33,4 +37,16 @@ internal interface DataModule {
     fun bindTimeTableDataSource(
         impl: TimeTableDataSourceImpl,
     ): TimeTableDataSource
+
+    @Singleton
+    @Binds
+    fun bindLocalOptionDataSource(
+        impl: LocalOptionDataSourceImpl,
+    ): LocalOptionDataSource
+
+    @Singleton
+    @Binds
+    fun bindLocalUserDataSource(
+        impl: LocalUserDataSourceImpl,
+    ): LocalUserDataSource
 }
